Moved code to not have a duplicate function, and fixed faker.factory logging
