+++ src/datadoc/frontend/fields/display_variables.py
@@ -15,6 +15,7 @@ from datadoc.frontend.fields.display_base import MetadataDropdownField
 from datadoc.frontend.fields.display_base import MetadataInputField
 from datadoc.frontend.fields.display_base import MetadataMultiLanguageField
 from datadoc.frontend.fields.display_base import MetadataPeriodField
+from datadoc.frontend.fields.display_base import get_data_source_options
 from datadoc.frontend.fields.display_base import get_enum_options
 
 
@@ -26,19 +27,6 @@ def get_measurement_unit_options() -> list[dict[str, str]]:
             "id": measurement_unit.code,
         }
         for measurement_unit in state.measurement_units.classifications
-    ]
-    dropdown_options.insert(0, {"title": "", "id": ""})
-    return dropdown_options
-
-
-def get_data_source_options() -> list[dict[str, str]]:
-    """Collect the unit type options."""
-    dropdown_options = [
-        {
-            "title": data_sources.get_title(enums.SupportedLanguages.NORSK_BOKMÅL),
-            "id": data_sources.code,
-        }
-        for data_sources in state.data_sources.classifications
     ]
     dropdown_options.insert(0, {"title": "", "id": ""})
     return dropdown_options
