+++ src/datadoc/logging_configuration/logging_config.py
@@ -52,3 +52,4 @@ def configure_logging(config: dict[str, Any] | None = None) -> None:
         }
 
     logging.config.dictConfig(config)
+    logging.getLogger("faker").setLevel(logging.ERROR)
